pyproject.toml
[project]
name = "demand-forecast-env"
version = "1.0.0"
description = "Demand Forecast Adjuster — an OpenEnv environment for S&OP demand planning"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.1",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.20.0",
    "fastapi>=0.100.0",
]

[project.scripts]
server = "server.app:main"
